add some manually drop parameter to help with dropping in the right order
This is not necessary because Framed* don't do anything action on the
inner type when dropping. So there could be a dandling reference on it
it would not be too much of an issue.

Otherwise, for sanity check it is better to drop the object at the right
time: before dropping the owning object that were just wrapped in a Box